Sync schema
