+++ reconcile/gql_definitions/fragments/resource_limits_requirements.gql
@@ -1,0 +1,6 @@
+# qenerate: plugin=pydantic_v1
+
+fragment ResourceLimitsRequirements on ResourceLimitsRequirements_v1 {
+    cpu
+    memory
+}
