+++ reconcile/gql_definitions/fragments/resource_requests_requirements.gql
@@ -1,0 +1,6 @@
+# qenerate: plugin=pydantic_v1
+
+fragment ResourceRequestsRequirements on ResourceRequestsRequirements_v1 {
+    cpu
+    memory
+}
